requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
faiss-cpu==1.7.4   # Use faiss-gpu for GPU support
numpy==1.26.0
sentence-transformers==2.2.2
torch==2.1.0       # For CUDA support, install appropriate version
transformers==4.34.1
nltk

